Atualizando para a API v3.0

+++ ClimaWeather/ClimaWeather/Converters/SomenteInteiroConverter.cs
@@ -6,7 +6,7 @@ namespace ClimaWeather.Converters {
 
     public class SomenteInteiroConverter : IValueConverter {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture) {
-            var valor = (decimal)value;
+            var valor = (double)value;
             var parteInteira = Math.Truncate(valor);
             return parteInteira.ToString();
         }
